+++ Misc/NEWS.d/next/C API/2021-03-23-20-53-41.bpo-43244.VK3sLH.rst
@@ -1,0 +1,11 @@
+Remove the compiler functions using ``struct _mod`` type, because the public
+AST C API was removed:
+
+* ``PyAST_Compile()``
+* ``PyAST_CompileEx()``
+* ``PyAST_CompileObject()``
+* ``PyFuture_FromAST()``
+* ``PyFuture_FromASTObject()``
+
+These functions were undocumented and excluded from the limited C API.
+Patch by Victor Stinner.
